refactor(agent-pane): Phase 3 Step 1 — scope layout NodesChanged to the virtualized partition
The layout-slice feeder dispatched NodesChanged with the FULL document.
Harmless in Phase 2 (slice is shadow; TanStack owns positions) but fatal
for Phase 3: positions() would prefix-sum the streaming-buffer nodes too.
Move the feeder from agent-view into AgentDocumentVirtualList (which owns
partition()) and scope NodesChanged/EstimateSet/ExpansionResolved to
partition().virtualizedNodes only — the region the slice models. Still
shadow; TanStack still renders, so zero user-visible change.

Adds docs/plans/PHASE3_LAYOUT_REDUCER_IMPL_PLAN.md (the plan this is step 1 of).
